fix: bring back js injection on `OnReceivedTitle`
