+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebView.kt
@@ -37,8 +37,8 @@ internal class MiniAppWebView(
         settings.allowUniversalAccessFromFileURLs = true
         settings.domStorageEnabled = true
         settings.databaseEnabled = true
-        webViewClient =
-            MiniAppWebViewClient(context, getWebViewAssetLoader(), customDomain, customScheme)
+        webViewClient = MiniAppWebViewClient(getWebViewAssetLoader(), customDomain, customScheme)
+        webChromeClient = MiniAppWebChromeClient(context)
 
         loadUrl(getLoadUrl())
     }
