+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebViewClient.kt
@@ -1,7 +1,5 @@
 package com.rakuten.tech.mobile.miniapp.display
 
-import android.content.Context
-import android.graphics.Bitmap
 import android.webkit.WebResourceRequest
 import android.webkit.WebResourceResponse
 import android.webkit.WebView
@@ -9,33 +7,17 @@ import android.webkit.WebViewClient
 import android.webkit.WebResourceError
 import androidx.annotation.VisibleForTesting
 import androidx.webkit.WebViewAssetLoader
-import java.io.BufferedReader
 
 internal class MiniAppWebViewClient(
-    context: Context,
     @VisibleForTesting internal val loader: WebViewAssetLoader,
     private val customDomain: String,
     private val customScheme: String
 ) : WebViewClient() {
 
-    @Suppress("TooGenericExceptionCaught", "SwallowedException")
-    @VisibleForTesting
-    internal val bridgeJs = try {
-        val inputStream = context.assets.open("bridge.js")
-        inputStream.bufferedReader().use(BufferedReader::readText)
-    } catch (e: Exception) {
-        null
-    }
-
     override fun shouldInterceptRequest(
         view: WebView,
         request: WebResourceRequest
     ): WebResourceResponse? = loader.shouldInterceptRequest(request.url)
-
-    override fun onPageStarted(webView: WebView, url: String?, favicon: Bitmap?) {
-        super.onPageStarted(webView, url, favicon)
-        webView.evaluateJavascript(bridgeJs) {}
-    }
 
     override fun onReceivedError(
         view: WebView,
